backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
langchain==0.1.0
langgraph==0.0.20
pydantic==2.5.2
reportlab==4.0.7
numpy==1.24.3
opencv-python==4.8.1.78
requests==2.31.0
python-dotenv==1.0.0
httpx==0.25.2

